osha-chatbot/requirements.txt
llama-index-core=0.12.2
llama-index-embeddings-huggingface = 0.4.0
llama-index-readers-file = 0.4.0
python-dotenv>=1.0.0
pydantic-settings>=2.1.0
arize-phoenix==5.12.0
arize-phoenix-otel = 0.6.1
llama_index.llms.bedrock = 0.3.1
openinference-instrumentation-bedrock = 0.1.12
openinference.instrumentation.llama_index = 3.0.4
openinference-instrumentation = 0.1.19
boto3 
botocore
arize = 7.29.2
editdistance #this is for experiments